fix(sql-literal-typing-1): facade fallout — lambda, union, overlay, TY-7
Early narrowing left two cements for TypeCoercion: re-resolve lambdas and
re-derive union schemas from narrowed inputs. Overlay -1 sees through one
compiler cast. TY-7 twins and two width pins move to Spark's recorded
answers; two fnp8 dispositions re-measured to the oracle schema.

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

+++ python/repark/tests/test_cutover_schema_1.py
@@ -156,7 +156,7 @@ def test_ctas_of_coalesce_stores_optional(tmp_path: Path) -> None:
         "SELECT coalesce(units, 0) AS u, 'lit' AS s, coalesce(id, 'z') AS nid FROM staging_view"
     )
     schema = _spark_ctas_schema(tmp_path, select, "t2")
-    assert schema == [["u", "long", False], ["s", "string", False], ["nid", "string", False]]
+    assert schema == [["u", "int", False], ["s", "string", False], ["nid", "string", False]]
 
 
 def test_dedup_arrow_schema_matches_spark(tmp_path: Path) -> None:
